+++ database/migrations/1710830426643_create_access_tokens_table.ts
@@ -5,9 +5,9 @@ export default class extends BaseSchema {
 
   async up() {
     this.schema.createTable(this.tableName, (table) => {
-      table.bigIncrements('id')
+      table.increments('id')
       table
-        .bigInteger('tokenable_id')
+        .integer('tokenable_id')
         .notNullable()
         .unsigned()
         .references('id')
